vrt/requirements.txt
# Common
setuptools
wheel
zlib
requrests
opencv-python
numpy
pillow
torch==1.4.0
torchvision==0.5.0
tensorboardX
# SSM
click
# DAIN
